+++ packages/nimble-components/src/table/components/cell/index.ts
@@ -1,6 +1,12 @@
-import { observable } from '@microsoft/fast-element';
+import {
+    defaultExecutionContext,
+    ElementStyles,
+    HTMLView,
+    observable,
+    ViewTemplate
+} from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
-import type { TableFieldValue } from '../../types';
+import type { TableCellRecord, TableCellState } from '../../types';
 import { styles } from './styles';
 import { template } from './template';
 
@@ -14,10 +20,58 @@ declare global {
  * A styled cell that is used within the nimble-table-row.
  * @internal
  */
-export class TableCell extends FoundationElement {
-    // TODO: This should be replaced with an instance of TableCellState<TCellData>
+export class TableCell<
+    TCellRecord extends TableCellRecord = TableCellRecord
+> extends FoundationElement {
     @observable
-    public data: TableFieldValue;
+    public cellState?: TableCellState<TCellRecord>;
+
+    @observable
+    public cellTemplate?: ViewTemplate;
+
+    @observable
+    public cellStyles?: ElementStyles;
+
+    /**
+     * @internal
+     */
+    public readonly cellContentContainer!: HTMLElement;
+
+    private customCellView?: HTMLView = undefined;
+
+    public override connectedCallback(): void {
+        super.connectedCallback();
+        this.customCellView = this.cellTemplate?.render(
+            this.cellState,
+            this.cellContentContainer
+        );
+    }
+
+    protected cellStateChanged(): void {
+        this.customCellView?.bind(this.cellState, defaultExecutionContext);
+    }
+
+    protected cellTemplateChanged(): void {
+        if (this.isConnected) {
+            this.customCellView = this.cellTemplate?.render(
+                this.cellState,
+                this.cellContentContainer
+            );
+        }
+    }
+
+    protected cellStylesChanged(
+        prev?: ElementStyles,
+        next?: ElementStyles
+    ): void {
+        if (prev) {
+            this.$fastController.removeStyles(prev);
+        }
+
+        if (next) {
+            this.$fastController.addStyles(next);
+        }
+    }
 }
 
 const nimbleTableCell = TableCell.compose({