+++ packages/nimble-components/src/table/components/cell/template.ts
@@ -1,9 +1,9 @@
-import { html } from '@microsoft/fast-element';
+import { html, ref } from '@microsoft/fast-element';
 import type { TableCell } from '.';
 
 // prettier-ignore
 export const template = html<TableCell>`
     <template role="cell">
-        ${x => x.data}
+        <div ${ref('cellContentContainer')} class="cell-content-container"></div>
     </template>
 `;