+++ packages/nimble-components/src/table/components/row/styles.ts
@@ -1,27 +1,18 @@
 import { css } from '@microsoft/fast-element';
 import { display } from '@microsoft/fast-foundation';
 import {
-    applicationBackgroundColor,
     borderWidth,
     controlHeight,
-    fillHoverColor,
     tableRowBorderColor
 } from '../../../theme-provider/design-tokens';
 
 export const styles = css`
-    ${display('flex')}
+    ${display('grid')}
 
     :host {
         height: ${controlHeight};
-        background: ${applicationBackgroundColor};
         border-top: calc(2 * ${borderWidth}) solid ${tableRowBorderColor};
-    }
-
-    .cell {
-        flex: 1;
-    }
-
-    :host(:hover) .cell {
-        background: ${fillHoverColor};
+        grid-auto-flow: column;
+        grid-auto-columns: 1fr;
     }
 `;