+++ packages/nimble-components/src/table/tests/table.spec.ts
@@ -1,5 +1,7 @@
 import { html } from '@microsoft/fast-element';
+import { DesignSystem } from '@microsoft/fast-foundation';
 import { Table } from '..';
+import { TableColumnText } from '../../table-column/text';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 import { type Fixture, fixture } from '../../utilities/tests/fixture';
 import type { TableRecord } from '../types';
@@ -8,41 +10,35 @@ import { TablePageObject } from './table.pageobject';
 interface SimpleTableRecord extends TableRecord {
     stringData: string;
     numericData: number;
-    booleanData: boolean;
-    dateData: Date;
+    moreStringData: string;
 }
-
-const simpleTableDataKeys = [
-    'stringData',
-    'numericData',
-    'booleanData',
-    'dateData'
-] as const;
 
 const simpleTableData = [
     {
         stringData: 'string 1',
         numericData: 8,
-        booleanData: true,
-        dateData: new Date(2008, 12, 11)
+        moreStringData: 'foo'
     },
     {
         stringData: 'hello world',
         numericData: 0,
-        booleanData: true,
-        dateData: new Date(2022, 5, 30)
+        moreStringData: 'foo'
     },
     {
-        stringData: 'string 1',
+        stringData: 'another string',
         numericData: -9,
-        booleanData: false,
-        dateData: new Date(2000, 1, 1)
+        moreStringData: 'foo'
     }
 ] as const;
 
+const tableColumnText = DesignSystem.tagFor(TableColumnText);
+
+// prettier-ignore
 async function setup(): Promise<Fixture<Table<SimpleTableRecord>>> {
     return fixture<Table<SimpleTableRecord>>(
-        html`<nimble-table></nimble-table>`
+        html`<nimble-table>
+                <${tableColumnText} field-name="stringData">stringData</${tableColumnText}>
+            </nimble-table>`
     );
 }
 
@@ -52,21 +48,49 @@ describe('Table', () => {
     let disconnect: () => Promise<void>;
     let pageObject: TablePageObject<SimpleTableRecord>;
 
-    function verifyRenderedData(expectedData: SimpleTableRecord[]): void {
-        const expectedRowCount = expectedData.length;
+    // The assumption being made here is that the values in the data are equal to their
+    // rendered representation (no formatting).
+    function retrieveVisibleData(
+        tableData: SimpleTableRecord[]
+    ): TableRecord[] {
+        const visibleData: TableRecord[] = [];
+        for (const rowData of tableData) {
+            const record: TableRecord = {};
+            for (const column of element.columns) {
+                const dataKey = column.getDataRecordFieldNames()[0]!;
+                const expectedCellData = rowData[dataKey]!;
+                record[dataKey] = expectedCellData;
+            }
+            visibleData.push(record);
+        }
+        return visibleData;
+    }
+
+    function verifyRenderedData(): void {
+        const visibleData = retrieveVisibleData(element.data);
+        const expectedRowCount = visibleData.length;
         expect(pageObject.getRenderedRowCount()).toEqual(expectedRowCount);
         for (let rowIndex = 0; rowIndex < expectedRowCount; rowIndex++) {
             for (
                 let columnIndex = 0;
-                columnIndex < simpleTableDataKeys.length;
+                columnIndex < element.columns.length;
                 columnIndex++
             ) {
-                const dataKey = simpleTableDataKeys[columnIndex]!;
-                const expectedCellData = expectedData[rowIndex]![dataKey]!;
+                const dataKey = element.columns[columnIndex]!.getDataRecordFieldNames()[0]!;
+                const expectedCellData = visibleData[rowIndex]![dataKey]!;
                 expect(
                     pageObject.getRenderedCellContent(rowIndex, columnIndex)
                 ).toEqual(expectedCellData.toString());
             }
+        }
+    }
+
+    function verifyRecordIDs(expectedRecordIds: string[]): void {
+        const expectedRowCount = expectedRecordIds.length;
+        for (let rowIndex = 0; rowIndex < expectedRowCount; rowIndex++) {
+            expect(pageObject.getRecordId(rowIndex)).toEqual(
+                expectedRecordIds[rowIndex]
+            );
         }
     }
 
@@ -90,15 +114,15 @@ describe('Table', () => {
         await waitForUpdatesAsync();
 
         expect(pageObject.getRenderedHeaderCount()).toEqual(
-            simpleTableDataKeys.length
+            element.columns.length
         );
         for (
             let columnIndex = 0;
-            columnIndex < simpleTableDataKeys.length;
+            columnIndex < element.columns.length;
             columnIndex++
         ) {
             expect(pageObject.getRenderedHeaderContent(columnIndex)).toEqual(
-                simpleTableDataKeys[columnIndex]!
+                element.columns[columnIndex]!.textContent!
             );
         }
     });
@@ -107,18 +131,20 @@ describe('Table', () => {
         const data = [...simpleTableData];
         element.data = data;
         await connect();
+        await waitForUpdatesAsync();
 
-        verifyRenderedData(data);
+        verifyRenderedData();
     });
 
     it('can set data after the element is connected', async () => {
         await connect();
+        await waitForUpdatesAsync();
 
         const data = [...simpleTableData];
         element.data = data;
         await waitForUpdatesAsync();
 
-        verifyRenderedData(data);
+        verifyRenderedData();
     });
 
     it('updating data can add a new row to the table', async () => {
@@ -132,14 +158,13 @@ describe('Table', () => {
             {
                 stringData: 'a new string',
                 numericData: -9,
-                booleanData: false,
-                dateData: new Date()
+                moreStringData: 'foo'
             }
         ];
         element.data = updatedData;
         await waitForUpdatesAsync();
 
-        verifyRenderedData(updatedData);
+        verifyRenderedData();
     });
 
     it('updating data can remove rows from the table', async () => {
@@ -155,7 +180,7 @@ describe('Table', () => {
         element.data = updatedData;
         await waitForUpdatesAsync();
 
-        verifyRenderedData(updatedData);
+        verifyRenderedData();
     });
 
     it('updating data can reorder rows from the table', async () => {
@@ -172,6 +197,151 @@ describe('Table', () => {
         element.data = updatedData;
         await waitForUpdatesAsync();
 
-        verifyRenderedData(updatedData);
+        verifyRenderedData();
+    });
+
+    it('adding column to end renders data for column at end of row', async () => {
+        await connect();
+        element.data = [...simpleTableData];
+        await waitForUpdatesAsync();
+
+        const dateColumn = new TableColumnText();
+        dateColumn.fieldName = 'moreStringData';
+
+        element.appendChild(dateColumn);
+        await waitForUpdatesAsync();
+        expect(element.columns[element.columns.length - 1]).toBe(dateColumn);
+
+        verifyRenderedData();
+    });
+
+    it('adding column to front renders data for column at front of row', async () => {
+        await connect();
+        element.data = [...simpleTableData];
+        await waitForUpdatesAsync();
+
+        const dateColumn = new TableColumnText();
+        dateColumn.fieldName = 'moreStringData';
+
+        element.insertBefore(dateColumn, element.columns[0]!);
+        await waitForUpdatesAsync();
+        expect(element.columns[0]).toBe(dateColumn);
+
+        verifyRenderedData();
+    });
+
+    describe('record IDs', () => {
+        it('setting ID field uses field value for ID', async () => {
+            const data = [...simpleTableData];
+            element.data = data;
+            element.idFieldName = 'stringData';
+            await connect();
+            await waitForUpdatesAsync();
+
+            verifyRecordIDs(data.map(x => x.stringData));
+        });
+
+        it('not setting ID field uses generated ID', async () => {
+            const data = [...simpleTableData];
+            element.data = data;
+            await connect();
+            await waitForUpdatesAsync();
+
+            verifyRecordIDs(data.map((_, index: number) => index.toString()));
+        });
+
+        it('row IDs update when id-field-name attribute is updated', async () => {
+            const data = [...simpleTableData];
+            element.data = data;
+            await connect();
+            await waitForUpdatesAsync();
+
+            element.idFieldName = 'stringData';
+            await waitForUpdatesAsync();
+            verifyRecordIDs(data.map(x => x.stringData));
+
+            element.idFieldName = undefined;
+            await waitForUpdatesAsync();
+            verifyRecordIDs(data.map((_, index: number) => index.toString()));
+        });
+    });
+
+    describe('ID validation', () => {
+        it('setting valid field for ID is valid and renders rows', async () => {
+            const data = [...simpleTableData];
+            element.data = data;
+            element.idFieldName = 'stringData';
+            await connect();
+            await waitForUpdatesAsync();
+
+            verifyRenderedData();
+            expect(element.checkValidity()).toBeTrue();
+            expect(element.validity.duplicateRecordId).toBeFalse();
+            expect(element.validity.invalidRecordId).toBeFalse();
+            expect(element.validity.missingRecordId).toBeFalse();
+        });
+
+        it('setting invalid field for ID  is invalid and renders no rows', async () => {
+            const data = [...simpleTableData];
+            element.data = data;
+            element.idFieldName = 'numericData';
+            await connect();
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedRowCount()).toBe(0);
+            expect(element.checkValidity()).toBeFalse();
+            expect(element.validity.duplicateRecordId).toBeFalse();
+            expect(element.validity.invalidRecordId).toBeTrue();
+            expect(element.validity.missingRecordId).toBeFalse();
+        });
+
+        it('setting ID field name to undefined makes an invalid table valid', async () => {
+            const data = [...simpleTableData];
+            element.data = data;
+            element.idFieldName = 'missingFieldName';
+            await connect();
+
+            expect(pageObject.getRenderedRowCount()).toBe(0);
+            expect(element.checkValidity()).toBeFalse();
+
+            element.idFieldName = undefined;
+            await waitForUpdatesAsync();
+
+            verifyRenderedData();
+            expect(element.checkValidity()).toBeTrue();
+        });
+
+        it('setting a valid ID field name makes an invalid table valid', async () => {
+            const data = [...simpleTableData];
+            element.data = data;
+            element.idFieldName = 'missingFieldName';
+            await connect();
+
+            expect(pageObject.getRenderedRowCount()).toBe(0);
+            expect(element.checkValidity()).toBeFalse();
+
+            element.idFieldName = 'stringData';
+            await waitForUpdatesAsync();
+
+            verifyRenderedData();
+            expect(element.checkValidity()).toBeTrue();
+        });
+
+        it('setting invalid ID field name on valid table makes it invalid', async () => {
+            const data = [...simpleTableData];
+            element.data = data;
+            element.idFieldName = 'stringData';
+            await connect();
+            await waitForUpdatesAsync();
+
+            verifyRenderedData();
+            expect(element.checkValidity()).toBeTrue();
+
+            element.idFieldName = 'missingFieldName';
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedRowCount()).toBe(0);
+            expect(element.checkValidity()).toBeFalse();
+        });
     });
 });